[narwhal] Notify client about commit boundaries
It is optimal to form sui checkpoints on the narwhal commit boundary(some motivation for this in #5763).

To do this, narwhal needs to notify sui when this boundary is reached.

This commit adds `notify_commit_boundary` to the execution state for this.

Note that this commit adds `notify_commit_boundary` method, but, in fact, for now this method is called on the **certificate** boundary, not commit boundary. This will do for now(it means checkpoints will be created more frequently then needed). We will need to refactor consensus crate a little to provide an actual commit boundary, however this can be done a bit later, when we get bulk of checkpoint work done.